Support customized logger
