+++ pool_func.go
@@ -117,15 +117,16 @@ func NewPoolWithFunc(size int, pf func(interface{}), options ...Option) (*PoolWi
 		return nil, ErrLackPoolFunc
 	}
 
-	opts := new(Options)
-	for _, option := range options {
-		option(opts)
-	}
+	opts := loadOptions(options...)
 
 	if expiry := opts.ExpiryDuration; expiry < 0 {
 		return nil, ErrInvalidPoolExpiry
 	} else if expiry == 0 {
 		opts.ExpiryDuration = DefaultCleanIntervalTime
+	}
+
+	if opts.Logger == nil {
+		opts.Logger = defaultLogger
 	}
 
 	p := &PoolWithFunc{
