+++ worker.go
@@ -23,7 +23,6 @@
 package ants
 
 import (
-	"log"
 	"runtime"
 	"time"
 )
@@ -54,10 +53,10 @@ func (w *goWorker) run() {
 				if ph := w.pool.options.PanicHandler; ph != nil {
 					ph(p)
 				} else {
-					log.Printf("worker exits from a panic: %v\n", p)
+					w.pool.options.Logger.Printf("worker exits from a panic: %v\n", p)
 					var buf [4096]byte
 					n := runtime.Stack(buf[:], false)
-					log.Printf("worker exits from panic: %s\n", string(buf[:n]))
+					w.pool.options.Logger.Printf("worker exits from panic: %s\n", string(buf[:n]))
 				}
 			}
 		}()
