+++ worker_func.go
@@ -23,7 +23,6 @@
 package ants
 
 import (
-	"log"
 	"runtime"
 	"time"
 )
@@ -54,10 +53,10 @@ func (w *goWorkerWithFunc) run() {
 				if ph := w.pool.options.PanicHandler; ph != nil {
 					ph(p)
 				} else {
-					log.Printf("worker with func exits from a panic: %v\n", p)
+					w.pool.options.Logger.Printf("worker with func exits from a panic: %v\n", p)
 					var buf [4096]byte
 					n := runtime.Stack(buf[:], false)
-					log.Printf("worker with func exits from panic: %s\n", string(buf[:n]))
+					w.pool.options.Logger.Printf("worker with func exits from panic: %s\n", string(buf[:n]))
 				}
 			}
 		}()
